% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_word_containing_your_word.R
\name{random_word_containing_your_word}
\alias{random_word_containing_your_word}
\title{A random word (or words, if desired) containing the word provided.}
\usage{
random_word_containing_your_word(word, number = 1)
}
\arguments{
\item{word}{single input string}

\item{number}{number of random words to return}
}
\value{
character vector of length specified of words containing the word
}
\description{
A random word (or words, if desired) containing the word provided.
}
\examples{
random_word_containing_your_word("dog")
random_word_containing_your_word("that",3)
}
